feat: remove tinyfiledialogs in favor of SDL and imgui
