+++ engine/ui/Popups.h
@@ -14,20 +14,6 @@ enum PopupFlags {
     POPUP_BUTTONS_RIGHT_TO_LEFT = 0x00000100,
 };
 
-/// Returns an absolute path to the file selected, or an empty string if none selected.
-std::string openFile(std::string_view pattern, std::string_view description = "");
-/// Returns an absolute path to the folder selected, or an empty string if none selected.
-std::string openFolder();
-/// Returns a relative path to the resource selected, or an empty string if none selected.
-std::string openResource(std::string_view pattern, std::string_view description = "");
-/// Return an absolute path to the file to create, or an empty string if Cancel pressed.
-std::string saveFile(std::string_view pattern, std::string_view defaultName = "", std::string_view description = "");
-
-/// Display a popup window with an input box.
-std::string input(std::string_view message = "", std::string_view title = "Input", std::string_view placeholder = "");
-/// Display a popup window with a password box.
-std::string password(std::string_view message = "", std::string_view title = "Password");
-
 /// Display a popup window with the specified message.
 void popup(std::string_view message, std::string_view title, unsigned int popupFlags, std::string_view ok = "OK");
 /// Display a popup info window with the specified message.
